added some useful utilities
Major functions added:

resample
—————
Resample or interpolate an array on to a grid with new extent.

make_cmap
——————
Make a color map compatible with matplotlib plotting functions.

close_polygon
———————
Return a list of points with the first point added to the end, simple
but very useful.

get_corners(x,y)
————————
Extract the corners of the data from its coordinates arrays x and y.

xy2pixel_coordinates
——————————
Get the pixel coordinates of xy points (can be lon,lat) on a grid with
`extent` and `shape`.